Rename stream_alert_cli_module to stream_alert_cli
