+++ stream_alert_cli/__init__.py
