+++ stream_alert_cli/version.py
@@ -15,7 +15,7 @@ from datetime import datetime
 import boto3
 from botocore.exceptions import ClientError
 
-from stream_alert_cli_module.logger import LOGGER_CLI
+from stream_alert_cli.logger import LOGGER_CLI
 
 
 class LambdaVersion(object):
