+++ tests/unit/stream_alert_cli/test_terraform_generate.py
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from stream_alert_cli_module import terraform_generate
+from stream_alert_cli import terraform_generate
 
 from nose.tools import assert_equal
